Squash #1897 - Update va profile id as well

+++ app/models.py
@@ -2176,7 +2176,7 @@ class VAProfileLocalCache(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     allowed = db.Column(db.Boolean, nullable=False)
-    va_profile_id = db.Column(db.Integer, nullable=False)
+    va_profile_id = db.Column(db.BigInteger, nullable=False)
     communication_item_id = db.Column(db.Integer, nullable=False)
     communication_channel_id = db.Column(db.Integer, nullable=False)
     source_datetime = db.Column(db.DateTime, nullable=False)
